entitys

+++ lab-06-orm/src/main/java/com/cydeo/entitys/Discount.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -16,7 +17,7 @@ import java.util.List;
 @Entity
 
 public class Discount extends BaseEntity{
-    private int discount;
+    private BigDecimal discount;
     @Enumerated(EnumType.STRING)
     private DiscountType discountType;
     private String name;
